Initial redesign of module functionality
