+++ core/src/loader/old/builtin_resolver.rs
