+++ core/src/loader/old/bundle.rs
