+++ core/src/loader/old/compile.rs
