+++ src/lib.rs
@@ -93,7 +93,7 @@
 pub use rquickjs_core::*;
 
 #[cfg(feature = "macro")]
-pub use rquickjs_macro::{bind, embed, FromJs, HasRefs, IntoJs, IntoJsByRef};
+pub use rquickjs_macro::{bind, FromJs, HasRefs, IntoJs, IntoJsByRef};
 
 // The following imports needed to linking docs
